refactor(cli): Move workspace to dedicated module
- Fix for https://github.com/markusz/buffrs/pull/3#discussion_r2419047381
